requirements.txt
appdirs==1.4.4
asgiref==3.8.1
autoslug==1.0.5
cffi==1.17.1
cryptography==44.0.2
diff-match-patch==20241021
Django==5.1.7
django-admin-autocomplete-filter==0.7.1
django-admin-list-filter-dropdown==1.0.3
django-autoslug==1.9.9
django-crispy-forms==2.4
django-extensions==3.2.3
django-import-export==4.3.7
django-iranian-cities==2.0.0
django-jalali==7.4.0
django-js-asset==3.1.2
django-phonenumber-field==8.0.0
django-ratelimit==4.1.0
django-redis==5.4.0
django-render-partial==0.4
django-richtextfield==1.6.2
django-timezone-field==7.1
fs==2.4.16
inflection==0.5.1
jalali_core==1.0.0
jdatetime==5.2.0
MarkupSafe==3.0.2
phonenumbers==9.0.1
pillow==11.1.0
pycparser==2.22
redis==5.2.1
setuptools==78.0.2
signals==0.0.2
six==1.17.0
sorl-thumbnail==12.11.0
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.1
unicode-slugify==0.1.5
Unidecode==1.3.8
Werkzeug==3.1.3
